scripts/LLaVA-OneVision-7B/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=3_kmp_llvm
accelerate=1.5.2=pyhd8ed1ab_0
aiohappyeyeballs=2.6.1=pyhd8ed1ab_0
aiohttp=3.11.13=py39h9399b63_0
aiosignal=1.3.2=pyhd8ed1ab_0
async-timeout=5.0.1=pyhd8ed1ab_1
attrs=25.3.0=pyh71513ae_0
aws-c-auth=0.7.22=h96bc93b_2
aws-c-cal=0.6.14=h88a6e22_1
aws-c-common=0.9.19=h4ab18f5_0
aws-c-compression=0.2.18=h83b837d_6
aws-c-event-stream=0.4.2=ha47c788_12
aws-c-http=0.8.1=h29d6fba_17
aws-c-io=0.14.8=h21d4f22_5
aws-c-mqtt=0.10.4=h759edc4_4
aws-c-s3=0.5.9=h594631b_3
aws-c-sdkutils=0.1.16=h83b837d_2
aws-checksums=0.1.18=h83b837d_6
aws-crt-cpp=0.26.9=he3a8b3b_0
aws-sdk-cpp=1.11.329=hba8bd5f_3
azure-core-cpp=1.13.0=h935415a_0
azure-identity-cpp=1.9.0=hd126650_0
azure-storage-blobs-cpp=12.13.0=h1d30c4a_0
azure-storage-common-cpp=12.8.0=ha3822c6_0
azure-storage-files-datalake-cpp=12.12.0=h0f25b8a_0
blas=1.0=mkl
brotli-python=1.1.0=py39hf88036b_2
bzip2=1.0.8=h4bc722e_7
c-ares=1.34.4=hb9d3cd8_0
ca-certificates=2025.1.31=hbcca054_0
certifi=2025.1.31=pyhd8ed1ab_0
cffi=1.17.1=py39h15c3d72_0
charset-normalizer=3.4.1=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_1
coloredlogs=15.0.1=pyhd8ed1ab_4
cuda-cudart=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.1=0
cuda-nvrtc=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.4.127=0
cuda-runtime=12.4.1=0
cuda-version=11.8=h70ddcb2_3
cudatoolkit=11.8.0=h4ba93d1_13
cudnn=9.8.0.87=cuda11.8
datasets=3.4.0=pyhd8ed1ab_0
dill=0.3.8=pyhd8ed1ab_0
einops=0.8.1=pyhd8ed1ab_0
ffmpeg=4.3=hf484d3e_0
filelock=3.18.0=pyhd8ed1ab_0
freetype=2.12.1=h267a509_2
frozenlist=1.5.0=py39h9399b63_1
fsspec=2024.12.0=pyhd8ed1ab_0
gflags=2.2.2=h5888daf_1005
giflib=5.2.2=hd590300_0
glog=0.7.1=hbabe93e_0
gmp=6.3.0=hac33072_2
gmpy2=2.1.5=py39h7196dd7_3
gnutls=3.6.13=h85f3911_1
h2=4.2.0=pyhd8ed1ab_0
hpack=4.1.0=pyhd8ed1ab_0
huggingface_hub=0.29.3=pyhd8ed1ab_0
humanfriendly=10.0=pyh707e725_8
hyperframe=6.1.0=pyhd8ed1ab_0
icu=73.2=h59595ed_0
idna=3.10=pyhd8ed1ab_1
jinja2=3.1.6=pyhd8ed1ab_0
jpeg=9e=h166bdaf_2
kernel-headers_linux-64=3.10.0=he073ed8_18
keyutils=1.6.1=h166bdaf_0
krb5=1.21.3=h659f571_0
lame=3.100=h166bdaf_1003
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.43=h712a8e2_4
lerc=4.0.0=h27087fc_0
libabseil=20240116.2=cxx17_he02047a_1
libarrow=16.1.0=hcb6531f_6_cpu
libarrow-acero=16.1.0=hac33072_6_cpu
libarrow-dataset=16.1.0=hac33072_6_cpu
libarrow-substrait=16.1.0=h7e0c224_6_cpu
libblas=3.9.0=16_linux64_mkl
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcblas=3.9.0=16_linux64_mkl
libcrc32c=1.1.2=h9c3ff4c_0
libcublas=12.4.5.8=0
libcufft=11.2.1.3=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcurl=8.8.0=hca28451_1
libcusolver=11.6.1.9=0
libcusparse=12.3.1.170=0
libdeflate=1.17=h0b41bf4_0
libedit=3.1.20250104=pl5321h7949ede_0
libev=4.33=hd590300_2
libevent=2.1.12=hf998b51_1
libffi=3.4.6=h2dba641_0
libgcc=14.2.0=h767d61c_2
libgcc-ng=14.2.0=h69a702a_2
libgfortran=14.2.0=h69a702a_2
libgfortran5=14.2.0=hf1ad2bd_2
libgomp=14.2.0=h767d61c_2
libgoogle-cloud=2.24.0=h2736e30_0
libgoogle-cloud-storage=2.24.0=h3d9a0c8_0
libgrpc=1.62.2=h15f2491_0
libhwloc=2.11.2=default_he43201b_1000
libiconv=1.18=h4ce23a2_1
libjpeg-turbo=2.0.0=h9bf148f_0
liblapack=3.9.0=16_linux64_mkl
liblzma=5.6.4=hb9d3cd8_0
liblzma-devel=5.6.4=hb9d3cd8_0
libnghttp2=1.58.0=h47da74e_1
libnpp=12.2.5.30=0
libnsl=2.0.1=hd590300_0
libnvfatbin=12.4.127=0
libnvjitlink=12.4.127=0
libnvjpeg=12.3.1.117=0
libopenblas=0.3.29=pthreads_h94d23a6_0
libopentelemetry-cpp=1.15.0=h66972aa_0
libopentelemetry-cpp-headers=1.15.0=ha770c72_0
libparquet=16.1.0=h6a7eafb_6_cpu
libpng=1.6.43=h2797004_0
libprotobuf=4.25.3=h08a7969_0
libre2-11=2023.09.01=h5a48ba9_2
libsqlite=3.46.0=hde9e2c9_0
libssh2=1.11.0=h0841786_0
libstdcxx=14.2.0=h8f9b012_2
libstdcxx-ng=14.2.0=h4852527_2
libthrift=0.19.0=hb90f79a_1
libtiff=4.5.0=h6adf6a1_2
libutf8proc=2.8.0=hf23e847_1
libuuid=2.38.1=h0b41bf4_0
libwebp=1.2.4=h1daa5a0_1
libwebp-base=1.2.4=h166bdaf_0
libxcb=1.13=h7f98852_1004
libxcrypt=4.4.36=hd590300_1
libxml2=2.12.7=hc051c1a_1
libzlib=1.2.13=h4ab18f5_6
llvm-openmp=15.0.7=h0cdce71_0
lz4-c=1.9.4=hcb278e6_0
markupsafe=3.0.2=py39h9399b63_1
mkl=2022.2.1=h84fe81f_16997
mpc=1.3.1=h24ddda3_1
mpfr=4.2.1=h90cbb55_3
mpmath=1.3.0=pyhd8ed1ab_1
multidict=6.1.0=py39h9399b63_2
multiprocess=0.70.16=py39h8cd3c5a_1
ncurses=6.5=h2d0b736_3
nettle=3.6=he412f7d_0
networkx=3.2.1=pyhd8ed1ab_0
nlohmann_json=3.11.3=he02047a_1
numpy=2.0.2=py39h9cb892a_1
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-cusparselt-cu12=0.6.2=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
onnx=1.16.2=py39hc348729_1
onnxruntime=1.19.2=py39hcbadb43_200_cuda
openh264=2.1.1=h780b84a_0
openjpeg=2.5.0=hfec8fc6_2
openssl=3.4.1=h7b32b05_0
orc=2.0.1=h17fec99_1
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=py39h3b40f6f_2
pillow=11.1.0=pypi_0
pip=25.0.1=pyh8b19718_0
prometheus-cpp=1.2.4=h90a924e_0
propcache=0.2.1=py39h9399b63_1
protobuf=4.25.3=py39hbb4dce6_1
psutil=7.0.0=py39h8cd3c5a_0
pthread-stubs=0.4=hb9d3cd8_1002
pyarrow=16.1.0=py39he8689d4_2
pyarrow-core=16.1.0=py39h80589db_2_cpu
pycparser=2.22=pyh29332c3_1
pysocks=1.7.1=pyha55dd90_7
python=3.9.18=h0755675_1_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python-flatbuffers=25.2.10=pyhbc23db3_0
python-tzdata=2025.1=pyhd8ed1ab_0
python-xxhash=3.5.0=py39h8cd3c5a_2
python_abi=3.9=5_cp39
pytorch=2.5.1=py3.9_cuda12.4_cudnn9.1.0_0
pytorch-cuda=12.4=hc786d27_7
pytorch-mutex=1.0=cuda
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.2=py39h9399b63_2
re2=2023.09.01=h7f4b329_2
readline=8.2=h8c095d6_2
regex=2024.11.6=py39h8cd3c5a_0
requests=2.32.3=pyhd8ed1ab_1
s2n=1.4.15=he19d79f_0
safetensors=0.5.3=py39he612d8f_0
setuptools=75.8.2=pyhff2d567_0
six=1.17.0=pyhd8ed1ab_0
snappy=1.2.1=h8bd8927_1
sympy=1.13.1=pypi_0
sysroot_linux-64=2.17=h0157908_18
tbb=2021.13.0=hceb3a55_1
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.21.1=py39he41eae5_0
torchaudio=2.5.1=py39_cu124
torchtriton=3.1.0=py39
torchvision=0.21.0=pypi_0
tqdm=4.67.1=pyhd8ed1ab_1
transformers=4.49.0=pyhd8ed1ab_0
triton=3.2.0=pypi_0
typing-extensions=4.12.2=hd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
tzdata=2025a=h78e105d_0
urllib3=2.3.0=pyhd8ed1ab_0
wheel=0.45.1=pyhd8ed1ab_1
xorg-libxau=1.0.12=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xxhash=0.8.3=hb9d3cd8_0
xz=5.6.4=hbcc6ac9_0
xz-gpl-tools=5.6.4=hbcc6ac9_0
xz-tools=5.6.4=hb9d3cd8_0
yaml=0.2.5=h7f98852_2
yarl=1.18.3=py39h9399b63_1
zlib=1.2.13=h4ab18f5_6
zstandard=0.23.0=py39h8cd3c5a_1
zstd=1.5.6=ha6fb4c9_0
